refactor: Edit diary setting module
