+++ feature/diary/src/main/java/com/teamoffroad/feature/diary/domain/repository/DiaryRepository.kt
@@ -8,11 +8,13 @@ import com.teamoffroad.feature.diary.domain.model.HexCode
 interface DiaryRepository {
     suspend fun getDiaryTutorialChecked(): Result<Boolean?>
     suspend fun patchDiaryTutorialChecked(): Result<Unit>
+    suspend fun patchDiaryCreateTime(diaryTime: Int)
     suspend fun getDiaryCreateTimeChecked(): Result<Boolean?>
     suspend fun patchDiaryCreateTimeChecked(): Result<Unit>
     suspend fun getDiaryFirstDate(): Result<DiaryFirstDate?>
     suspend fun getDiaryMonthlyHex(year: Int, month: Int): Result<Map<String, List<HexCode>>?>
     suspend fun getDiaryLatest(previousCount: Int): Result<DiaryLatest?>
+    suspend fun getDiaryCheckLatest(): Result<Boolean?>
     suspend fun getDiaryByDate(
         date: String,
         previousCount: Int,
