--- feature/mypage/src/main/java/com/teamoffroad/feature/mypage/data/repository/DiarySettingRepositoryImpl.kt
@@ -1,16 +1,0 @@
-package com.teamoffroad.feature.mypage.data.repository
-
-import com.teamoffroad.feature.mypage.data.remote.service.DiarySettingService
-import com.teamoffroad.feature.mypage.domain.repository.DiarySettingRepository
-import javax.inject.Inject
-
-class DiarySettingRepositoryImpl @Inject constructor(
-    private val diarySettingService: DiarySettingService,
-) : DiarySettingRepository {
-    override suspend fun patchDiaryCreateTime(diaryTime: Int) {
-        diarySettingService.patchDiaryCreateTime(diaryTime)
-    }
-
-    override suspend fun getDiaryCheckLatest(): Result<Boolean?> =
-        runCatching { diarySettingService.getDiaryCheckLatest().data?.doesNotExistOrChecked }
-}
