--- feature/mypage/src/main/java/com/teamoffroad/feature/mypage/domain/repository/DiarySettingRepository.kt
@@ -1,6 +1,0 @@
-package com.teamoffroad.feature.mypage.domain.repository
-
-interface DiarySettingRepository {
-    suspend fun patchDiaryCreateTime(diaryTime: Int)
-    suspend fun getDiaryCheckLatest(): Result<Boolean?>
-}
