--- feature/mypage/src/main/java/com/teamoffroad/feature/mypage/domain/usecase/GetDiaryCheckLatestUseCase.kt
@@ -1,9 +1,0 @@
-package com.teamoffroad.feature.mypage.domain.usecase
-
-import com.teamoffroad.feature.mypage.domain.repository.DiarySettingRepository
-
-class GetDiaryCheckLatestUseCase(
-    private val diarySettingRepository: DiarySettingRepository
-) {
-    suspend operator fun invoke() = diarySettingRepository.getDiaryCheckLatest()
-}
